+++ packages/cli/src/lib/errors/projectError.ts
@@ -1,0 +1,6 @@
+export class InvalidProjectError extends Error {
+  public static readonly message = 'INVALID_PROJECT_ERROR';
+  public constructor() {
+    super(InvalidProjectError.message);
+  }
+}