+++ packages/cli/src/lib/project/project.ts
@@ -1,31 +1,40 @@
-import {createWriteStream, existsSync, unlinkSync} from 'fs';
+import {createWriteStream, existsSync, unlinkSync, writeFileSync} from 'fs';
 import {join} from 'path';
 import {cli} from 'cli-ux';
 import archiver from 'archiver';
+import {InvalidProjectError} from '../errors';
+import extract from 'extract-zip';
 
 export class Project {
-  public constructor(private pathToProject: string) {
-    this.ensureProjectCompliance();
+  public constructor(private pathToProject: string) {}
+
+  public async refresh(projectContent: Blob) {
+    const buffer = await projectContent.arrayBuffer();
+    const view = new DataView(buffer);
+    writeFileSync(this.pathToTemporaryZip, view);
+    await extract(this.pathToTemporaryZip, {dir: this.resourcePath});
+    this.deleteTemporaryZipFile();
   }
 
   public deleteTemporaryZipFile() {
     unlinkSync(this.pathToTemporaryZip);
   }
 
-  public ensureProjectCompliance() {
+  private ensureProjectCompliance() {
     /*
      * TODO: CDX-354: add checks to ensure the project is indeed a valid project
      * e.g. * Check if path to resources is a folder
      *      * Check if the root has a valid config file
      */
 
-    if (!existsSync(this.pathToResources)) {
-      throw new Error('Invalid Project. TODO: better error message');
+    if (!existsSync(this.resourcePath)) {
+      throw new InvalidProjectError();
     }
   }
 
   public async compressResources() {
     try {
+      this.ensureProjectCompliance();
       await new Promise<void>((resolve, reject) => {
         const pathToTemporaryZip = this.pathToTemporaryZip;
         const outputStream = createWriteStream(pathToTemporaryZip);
@@ -35,7 +44,7 @@ export class Project {
         archive.on('error', (err) => reject(err));
 
         archive.pipe(outputStream);
-        archive.directory(this.pathToResources, false);
+        archive.directory(this.resourcePath, false);
         archive.finalize();
       });
       return this.pathToTemporaryZip;
@@ -48,7 +57,7 @@ export class Project {
     return join(this.pathToProject, 'snapshot.zip');
   }
 
-  private get pathToResources() {
+  private get resourcePath() {
     return join(this.pathToProject, 'resources');
   }
 }